Updated rouge function to use doubles over integers

+++ src/no/runsafe/framework/minecraft/event/entity/RunsafeEntityDamageEvent.java
@@ -13,7 +13,7 @@ public class RunsafeEntityDamageEvent extends RunsafeEntityEvent implements Canc
 		event = toWrap;
 	}
 
-	public int getDamage()
+	public double getDamage()
 	{
 		return event.getDamage();
 	}
